Put _all_ the Let notation functions in one collection
There is no reason to keep the `#()` and `#Assign()`
functions out just because they are in My Standard Library.
Duplicates are OK, since MSL is designed to be used
on a project startup as a "one copy, one paste" shop,
whereas the other items are more for "as needed", and
if there are whole collections, they should all be there.